feat!(api): move provider(s) endpoints to /controllers/providersController.js; rename for consistency to providers
